+++ packages/build-scripts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackwright/build-scripts
 
+## 0.2.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [6820928]
+  - @stackwright/types@0.3.2-alpha.0
+
 ## 0.2.1
 
 ### Patch Changes